backend/requirements.txt
asgiref==3.8.1
babel==2.17.0
charset-normalizer==3.4.2
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-money==3.5.3
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
pillow==11.3.0
psycopg2==2.9.10
py-moneyed==3.0
PyJWT==1.7.1
reportlab==4.4.2
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
